requirements.txt
pip~=25.0.1
mysql-connector-python~=9.2.0
setuptools~=76.0.0

-e .